links generation

+++ src/main/java/marchcat/pictures/LinkRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jdbc.repository.query.Modifying;
 import org.springframework.data.jdbc.repository.query.Query;
 import org.springframework.data.repository.CrudRepository;
 
-public interface LinkRepository extends CrudRepository<Link, Long> {
+public interface LinkRepository extends CrudRepository<Link, Integer> {
 
 	@Modifying
 	@Query("INSERT INTO links(id, link) VALUES(:id, :link)")
